ai-agent/requirements.txt
flask==3.1.0
anthropic==0.49.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
redis==5.2.1
bleach==6.2.0
PyJWT==2.10.1
gunicorn==23.0.0
pytest==8.3.4
pytest-mock==3.14.0
